+++ src/lib/hooks.ts
@@ -1,7 +1,10 @@
-import { useContext } from 'react';
+import { useContext, useEffect, useState } from 'react';
 
 import { AppContext } from '@/contexts/AppContextProvider';
 import { FeedbackItemsContext } from '@/contexts/FeedbackItemsContextProvider';
+
+import { URL } from '@/lib/constants';
+import { handleErrorStatuses } from '@/lib/handleErrors';
 
 const useAppContext = (componentName: string) => {
   const context = useContext(AppContext);
@@ -27,4 +30,49 @@ const useFeedbackItemsContext = (componentName: string) => {
   return context;
 };
 
-export { useAppContext, useFeedbackItemsContext };
+const useFeedbackItems = () => {
+  const [loading, setLoading] = useState(false);
+  const [errorMessage, setErrorMessage] = useState('');
+
+  const { feedbackItems, setFeedbackItems } = useAppContext(
+    'FeedbackItemsContextProvider',
+  );
+
+  const fetchData = async () => {
+    setLoading(true);
+
+    try {
+      const response = await fetch(URL);
+
+      if (response.ok) {
+        const data = await response.json();
+
+        setFeedbackItems(data.feedbacks);
+        setLoading(false);
+      } else {
+        const { status } = response;
+
+        handleErrorStatuses(status);
+      }
+    } catch (error) {
+      if (error instanceof Error) {
+        setErrorMessage(error.message);
+      }
+
+      setLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    fetchData();
+  }, []);
+
+  return {
+    loading,
+    errorMessage,
+    feedbackItems,
+    setFeedbackItems,
+  };
+};
+
+export { useAppContext, useFeedbackItemsContext, useFeedbackItems };
